+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fix *ConnPool.Deallocate() not deleting prepared statement from map
 * Fix stdlib not logging unprepared query SQL (Krzysztof Dryś)
+* Fix Rows.Values() with varchar binary format
 
 ## Features